requirements.txt

Flask==3.0.3
Flask-Cors==4.0.1
functions == 0.7.0
openai == 1.2.3
packaging == 23.2
requests == 2.31.0
python-dotenv==1.0.1
gunicorn == 22.0.0
